+++ Editor/McpProjectSettingsTool.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: d9fe2705694ffec47a2b3f32f6310508
