+++ Editor/McpProjectSettingsToolBuilder.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: fe5cc8487bf9ab648ae81544c14f8e0f
