+++ Runtime/McpProjectSettingsToolBuilder.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 12b4ebb321fe5454a8b602452a2abb65
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
